pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "jitter"
version = "0.1.0"
description = "An AI agent that discovers trending ideas and ships code daily"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.52.0",
    "tavily-python>=0.5.0",
    "PyGithub>=2.5.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.11.0",
    "click>=8.0",
    "ruamel.yaml>=0.17.0",
    "rich>=13.0",
    "tenacity>=9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.14.0",
]

[project.scripts]
jitter = "jitter.cli:cli"

[tool.setuptools.packages.find]
include = ["jitter*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
